requirements.txt
streamlit
whisper
PyPDF2
ffmpeg-python
pillow
fpdf
numpy
torch
datetime
openai
python-dotenv